+++ apps/frontend/app/(client)/(main)/course/[courseId]/_components/GradeAccordion.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import type { AssignmentGrade } from '@/app/(client)/_libs/apis/assignmentSubmission'
 import { assignmentSubmissionQueries } from '@/app/(client)/_libs/queries/assignmentSubmission'
 import {
   Accordion,
@@ -10,6 +9,7 @@ import {
 } from '@/components/shadcn/accordion'
 import { Dialog } from '@/components/shadcn/dialog'
 import { cn, convertToLetter, dateFormatter } from '@/libs/utils'
+import type { AssignmentGrade } from '@/types/type'
 import { useQuery } from '@tanstack/react-query'
 import { useState } from 'react'
 import { GradeDetailModal } from '../grade/_components/GradeDetailModal'
