+++ apps/frontend/types/type.ts
@@ -380,3 +380,31 @@ export interface CalendarAssignment {
   start: Date
   end: Date
 }
+
+export interface AssignmentGrade {
+  id: number
+  title: string
+  endTime: string
+  autoFinalizeScore: boolean
+  isFinalScoreVisible: boolean
+  isJudgeResultVisible: boolean
+  week: number
+  userAssignmentFinalScore: number | null
+  userAssignmentJudgeScore: number | null
+  assignmentPerfectScore: number
+  problems: ProblemGrade[]
+}
+export interface ProblemGrade {
+  id: number
+  title: string
+  order: number
+  maxScore: number
+  problemRecord: ProblemRecord | null
+}
+
+export interface ProblemRecord {
+  finalScore: number
+  score: number
+  isSubmitted: boolean
+  comment: string
+}
