+++ collection/client/Assignment/Get assignment by ID/Succeed.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/assignment/24
+  url: {{baseUrl}}/assignment/34
   body: none
   auth: none
 }
